feat: inline dir and value for all coords entries
