+++ src/components/RenderForm.js
@@ -3,8 +3,7 @@ import Form from '@rjsf/core';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import schema_map from '../utilities/constants';
 import { checkDraft2020, preProcessing } from '../utilities/schemaHandlers';
-import ObjectFieldTemplate from '../schemas/ui-schemas/ObjectFieldTemplate';
-import uiSchema from '../schemas/ui-schemas/ephys-session-ui-schema';
+import uiSchema from '../schemas/ui-schemas/EphysSessionUISchema';
 
 const Ajv2020 = require("ajv/dist/2020")
 
@@ -38,46 +37,7 @@ export default function RenderForm (props) {
     FileSaver.saveAs(blob, `${filename}.json`);
   };
 
-  //const uiSchema = require('../schemas/ui-schemas/ephys-session-ui-schema.js')
-  
-  /*const uiSchema = {
-    "probe_streams": {
-      "items": {
-            "probes": {
-              "items": {
-                    "targeted_ccf_coordinates": {
-                        "ui:options": {
-                            "orderable": false,
-                            "removable": false,
-                            "addable": false,
-                            //"inline": true
-                        },
-                        "items": { 
-                          "ui:ObjectFieldTemplate": ObjectFieldTemplate,
-                            "value": {
-                             "ui:placeholder": "0"
-                            }
-                        }
-                    },
-                    "targeted_lab_coordinates": {
-                        "ui:options": {
-                            "orderable": false,
-                            "removable": false,
-                            "addable": false
-                        },
-                        "items": {
-                            "value": {
-                             "ui:placeholder": "0"
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}*/
-
-
+  // TODO: uiSchema selection based on schema (prob should go in schemaHandlers)
   if(processedSchema){
     return (
       <Form schema={processedSchema}
